feat: Auth/User/Admin 컨트롤러 분리
대충 보니까..
AuthController - 도이(회원가입/로그인/인증인가만 담당)
UserController - 마이페이지, 유저의 활동 기반 정보들 담당자(민재?)
AdminController - 관리자가 하는 작업들 담당자(가연?)

같아요
