+++ src/main/java/com/b1a4/cafeOn/configs/SecurityConfig.java
@@ -44,6 +44,9 @@ public class SecurityConfig {
                 .authorizeHttpRequests(authorize -> authorize
                         .requestMatchers("/", "/api/auth/**", "/swagger-ui/**", "/v3/api-docs/**", "/api-docs", "/api-docs-json")  // 요청 경로가 일치하는 애들한테는
                         .permitAll()    // /, /api/auth/** 경로는 인증 안해도 되게 모두 허용하겠다!!(이코드 안쓰면 우리코드랑 관련없는 무슨 security 기본 로그인화면뜸)
+//                        ** URL 기반 인가 규칙 추가 **
+                        .requestMatchers("/api/admin/**").hasRole("ADMIN")  // 관리자 권한
+                        .requestMatchers("/api/user/**").hasAnyRole("USER", "ADMIN")    // 유저 권한
                         .anyRequest().authenticated()); // 그 이외의 모든 경로는 인증 해야됨
